pyproject.toml
[tool.poetry]
name = "pythondjangoproject"
version = "0.1.0"
description = "Django Rest Framework app to currency converte"
authors = ["Regiane <nascire@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"
pytest-factoryboy = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mocker = "^1.1.1"
pytest-django = "^4.5.2"
black = "^22.6.0"
flake8 = "^5.0.4"
pytest-mock = "^3.8.2"
pytest-freezegun = "^0.4.2"

[tool.black]
line-length = 99
target-version = ['py37']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
